feat: #8227 pr feedback
