+++ packages/payments-service/src/core/cors-header-interceptor.ts
@@ -2,12 +2,23 @@ import { CallHandler, ExecutionContext, Injectable, NestInterceptor } from '@nes
 import { Observable } from 'rxjs'
 import { Response } from 'express'
 
+const allowedOrigins = [
+  'https://payments.dev.paas.reapit.cloud',
+  'https://payments.prod.paas.reapit.cloud',
+  'https://payments-portal.dev.paas.reapit.cloud',
+  'https://payments-portal.prod.paas.reapit.cloud',
+]
+
+if (process.env.APP_ENV !== 'production') {
+  allowedOrigins.push('http://localhost:8080')
+}
+
 @Injectable()
 export class CorsHeaderInterceptor implements NestInterceptor {
   intercept(context: ExecutionContext, next: CallHandler): Observable<any> {
     const response = context.switchToHttp().getResponse<Response>()
 
-    response.setHeader('Access-Control-Allow-Origin', '*')
+    response.setHeader('Access-Control-Allow-Origin', allowedOrigins.join(','))
 
     return next.handle()
   }
